refactor leaderelector package move it under the k8s package
