--- leaderelector/leader_elector.go
@@ -1,7 +1,0 @@
-package leaderelector
-
-import "context"
-
-type LeaderElector interface {
-	Run(ctx context.Context)
-}
